2.6.36: Make features->num_contacts_left signed
Silence the following warning and ensure that existing comparisons of "<=0"
actually make complete sense.

2.6.36/wacom_wac.c:1217:21: warning: comparison of distinct pointer types lacks a cast [enabled by default]

+++ 2.6.36/wacom_wac.c
@@ -908,7 +908,7 @@ static int wacom_mt_touch(struct wacom_wac *wacom)
 	}
 
 	/* There are at most 5 contacts per packet */
-	contacts_to_send = min(5, (int)features->num_contacts_left);
+	contacts_to_send = min(5, features->num_contacts_left);
 
 	for (i = 0; i < contacts_to_send; i++) {
 		id = get_unaligned_le16(&data[k]);
